src/finance_section/requirements.txt
aiosignal
async-timeout
finvizfinance
functions-framework
google-cloud-bigquery
grpcio-status
httpx
pip-chill
praw
pytest
python-dotenv
streamlit
yarl
yfinance
